It has been discovered that whitelisted clients currently in the session can also tunnel other people into the session if they a) cannot connect to you directly and b) cannot get tunnelled through R* and maybe c) the peer's network allows such a thing to happen and maybe d) the person getting tunnelled has been in the session once before.
It appears that there's a lot that needs to happen for this to occur, but it still can happen and if it does your session integrity is gone.

So, I'm going to have a crack at solving this and see if my theory works. Testing this might be impossible if I can't consistently replicate the scenario.