I have noticed that Guardian appears to "leak" packets for several seconds before appearing to take action on a connection and drop packets. This is not a PyDivert bug; my private proof-of-concept does not have this issue, and is using almost the same logic in the filter loop (and is also using PyDivert). My PoC does not leak packets such that they are visible in Wireshark, and it's possible that client tunnelling was only possible because of this leak.
Will try to figure out why Guardian is leaking packets and make this solution air-tight, same as my PoC.

For reference, when using my PoC, players not on the whitelist do not even get the chance to attempt to connect to the session, but Guardian doesn't block for quite some time for some reason, allowing non-whitelisted clients to not only get session info, but also *technically* be in the session for a brief period of time before getting dropped.

Will clean up things once I discover the source of the bug.