Changed some session "stability tags" based on new testing.
Also, I've discovered that certain attributes of PyDivert packets cannot be pickled. The current error is relating to packets containing a MemoryView object which due to not being a finite object has no bounds blah blah blah can't be pickled.

Also learned that the pickle module isn't safe and code execution can occur, so the next thing to work on is a function which can sanitize PyDivert packets so they're' safer and can actually be pickled.