So, after a week of struggles, I've finally learned that PyDivert packets cannot be pickled / piped to other processes.
This is because PyDivert packets use a very specific data structure to increase performance and reduce memory, and this data-type is impossible to transfer directly.

Therefore, I will have to either construct a new object containing only the data necessary, or, pipe all the data over to construct the Packet individually, then actually construct the Packet in the new Process.